feat: move screensaver settings to android-specific source set and fix screensaver scaffold padding
- Move `ScreensaverTypeSettings` from common main (`GeneralSettings.kt` / `LayoutScreen.kt`) to android main (`CustomSettings.kt`), injecting `NewSettingsHandling` via Koin.
- Pass the scaffold's inner padding down to `SensorRotatedLayout` across screensaver screens (`CoverCarouselScreen`, `CustomListRotationScreen`, `HistoryFeedScreen`, and `StatsDashboardScreen`).
- Implement `SlowScroll` with a explicitly managed `LazyGridState` in `CustomListGrid`.
