+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/settings/behavior/LayoutScreen.kt
@@ -6,13 +6,11 @@ import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
 import com.programmersbox.datastore.NewSettingsHandling
-import com.programmersbox.kmpuiviews.presentation.components.item
 import com.programmersbox.kmpuiviews.presentation.components.settings.CategoryGroupListItem
 import com.programmersbox.kmpuiviews.presentation.settings.SettingsScaffold
 import com.programmersbox.kmpuiviews.presentation.settings.general.DetailPaneSettings
 import com.programmersbox.kmpuiviews.presentation.settings.general.GridTypeSettings
 import com.programmersbox.kmpuiviews.presentation.settings.general.NavigationBarSettings
-import com.programmersbox.kmpuiviews.presentation.settings.general.ScreensaverTypeSettings
 import com.programmersbox.kmpuiviews.utils.ComposeSettingsDsl
 import org.koin.compose.koinInject
 
@@ -34,10 +32,6 @@ fun LayoutScreen(
 
         CategoryGroupListItem {
             item { NavigationBarSettings(handling = handling) }
-        }
-
-        CategoryGroupListItem {
-            item { ScreensaverTypeSettings(handling = handling) }
         }
 
         composeSettingsDsl.layoutSettings()
